GDB-11741: Add Query Aborting Functionality
## What
The GraphQL playground has been extended with the ability to cancel a running query and notify about the cancellation.
Previously, there was an implementation that "canceled" queries, but it did not actually cancel the request—it only ignored the result. Additionally, there was no notification about the event.

## Why
We need to properly cancel the request and notify the GraphQL playground client when this event occurs so they can take appropriate actions after cancellation.

## How
- The API uses fetch for requests, so I extended the request handling with AbortController, which allows queries to be canceled.
- The CreateFetcherOptions interface was extended with a new property, onAbortQuery, a function that is called when a query is canceled.

### Additional Work
The headers property of CreateFetcherOptions has been extended to support functions as well. If a function is provided, it will be called and must return an object containing headers that will be added to the request.